+++ CHANGELOG.md
@@ -7,8 +7,13 @@ See [version roadmap](https://www.graalvm.org/release-notes/version-roadmap/) fo
 
 ## Version 22.2.0
 * GraalVM JavaScript is now an installable component of GraalVM. It can be installed with `gu install js`.
+* Enabled option `js.foreign-object-prototype` by default. Polyglot Interop objects now get a fitting JavaScript prototype assigned unless explicitly turned off using this flag.
+* Added intermediate prototype for foreign objects to simplify adapting functionality.
+* Removed deprecated experimental option `experimental-foreign-object-prototype`.
 * Removed experimental option `commonjs-global-properties`. The same functionality can be achieved in user code with a direct call to `require()` after context creation.
 * Added an experimental option `--js.zone-rules-based-time-zones` that allows to use timezone-related data from `ZoneRulesProvider` (instead of ICU4J data files).
+* Temporal objects can be converted to compatible Java objects when possible, using the `Value` API's methods like `asDate()`.
+
 
 ## Version 22.1.0
 * Updated Node.js to version 16.14.2.